Refactor websocket client lifecycle helpers
