+++ custom_components/termoweb/ws_shared.py
@@ -1,0 +1,173 @@
+"""Shared helpers for TermoWeb websocket clients."""
+
+from __future__ import annotations
+
+import asyncio
+from collections.abc import Iterable
+import contextlib
+from dataclasses import dataclass
+import logging
+import time
+from typing import Any
+
+import aiohttp
+from homeassistant.core import HomeAssistant
+from homeassistant.helpers.dispatcher import async_dispatcher_send
+
+from .const import DOMAIN, signal_ws_status
+
+_LOGGER = logging.getLogger(__name__)
+
+
+class HandshakeError(RuntimeError):
+    """Raised when the websocket handshake fails."""
+
+    def __init__(self, status: int, url: str, body_snippet: str) -> None:
+        """Store response information for logging and retries."""
+
+        super().__init__(f"handshake failed (status={status})")
+        self.status = status
+        self.url = url
+        self.body_snippet = body_snippet
+
+
+@dataclass
+class WSStats:
+    """Simple counter bucket for websocket activity."""
+
+    frames_total: int = 0
+    events_total: int = 0
+    last_event_ts: float = 0.0
+    last_paths: list[str] | None = None
+
+
+class TermoWebWSShared:
+    """Mixin providing lifecycle and status helpers for websocket clients."""
+
+    def __init__(
+        self,
+        hass: HomeAssistant,
+        *,
+        entry_id: str,
+        dev_id: str,
+        task_name: str | None = None,
+    ) -> None:
+        """Initialize shared websocket bookkeeping fields."""
+
+        self.hass = hass
+        self.entry_id = entry_id
+        self.dev_id = dev_id
+        self._runner_task_name_value = task_name
+        self._task: asyncio.Task | None = None
+        self._ws: aiohttp.ClientWebSocketResponse | None = None
+        self._hb_task: asyncio.Task | None = None
+        self._closing = False
+        self._connected_since: float | None = None
+        self._healthy_since: float | None = None
+        self._stats = WSStats()
+
+    # ----------------- Public control -----------------
+
+    def start(self) -> asyncio.Task:
+        """Start the websocket runner task if not already running."""
+
+        if self._task and not self._task.done():
+            return self._task
+        self._closing = False
+        name = self._runner_task_name()
+        self._task = self.hass.loop.create_task(self._runner(), name=name)
+        return self._task
+
+    async def stop(self) -> None:
+        """Stop the websocket client and cancel background tasks."""
+
+        self._closing = True
+        if self._hb_task:
+            self._hb_task.cancel()
+            with contextlib.suppress(asyncio.CancelledError):
+                await self._hb_task
+            self._hb_task = None
+
+        if self._ws:
+            with contextlib.suppress(*self._ws_close_exceptions()):
+                await self._ws.close(
+                    code=aiohttp.WSCloseCode.GOING_AWAY, message=b"client stop"
+                )
+            self._ws = None
+
+        if self._task:
+            self._task.cancel()
+            with contextlib.suppress(asyncio.CancelledError):
+                await self._task
+            self._task = None
+
+        self._update_status("stopped")
+
+    def is_running(self) -> bool:
+        """Return True if the websocket runner task is active."""
+
+        return bool(self._task and not self._task.done())
+
+    # ----------------- Helpers -----------------
+
+    def _runner_task_name(self) -> str:
+        return self._runner_task_name_value or f"{DOMAIN}-ws-{self.dev_id}"
+
+    def _ws_close_exceptions(self) -> Iterable[type[BaseException]]:
+        return (TimeoutError, aiohttp.ClientError, RuntimeError)
+
+    def _update_status(self, status: str) -> None:
+        """Update shared websocket state and dispatch notifications."""
+
+        state_bucket = self.hass.data[DOMAIN][self.entry_id].setdefault("ws_state", {})
+        state = state_bucket.setdefault(self.dev_id, {})
+        now = time.time()
+        state["status"] = status
+        state["last_event_at"] = self._stats.last_event_ts or None
+        state["healthy_since"] = self._healthy_since
+        state["healthy_minutes"] = (
+            int((now - self._healthy_since) / 60) if self._healthy_since else 0
+        )
+        state["frames_total"] = self._stats.frames_total
+        state["events_total"] = self._stats.events_total
+
+        async_dispatcher_send(
+            self.hass,
+            signal_ws_status(self.entry_id),
+            {"dev_id": self.dev_id, "status": status},
+        )
+
+    def _mark_event(self, *, paths: list[str] | None) -> None:
+        """Record websocket activity and update health state."""
+
+        now = time.time()
+        self._stats.last_event_ts = now
+        if paths:
+            self._stats.events_total += 1
+            if _LOGGER.isEnabledFor(logging.DEBUG):
+                uniq: list[str] = []
+                for path in paths:
+                    if path not in uniq:
+                        uniq.append(path)
+                    if len(uniq) >= 5:
+                        break
+                self._stats.last_paths = uniq
+
+        domain_bucket: dict[str, Any] = self.hass.data.setdefault(DOMAIN, {})
+        entry_bucket: dict[str, Any] = domain_bucket.setdefault(self.entry_id, {})
+        state_bucket: dict[str, dict[str, Any]] = entry_bucket.setdefault(
+            "ws_state", {}
+        )
+        state: dict[str, Any] = state_bucket.setdefault(self.dev_id, {})
+        state["last_event_at"] = now
+        state["frames_total"] = self._stats.frames_total
+        state["events_total"] = self._stats.events_total
+
+        if (
+            self._connected_since
+            and not self._healthy_since
+            and (now - self._connected_since) >= 300
+        ):
+            self._healthy_since = now
+            self._update_status("healthy")
+
